% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dtm.R
\name{dtm_melter}
\alias{dtm_melter}
\title{Melt a DTM into a triplet data frame}
\usage{
dtm_melter(dtm)
}
\arguments{
\item{dtm}{Document-term matrix with terms as columns. Works with DTMs
produced by any popular text analysis package, or using the
\code{dtm_builder()} function.}
}
\value{
returns data frame with three columns
}
\description{
Takes a DTM converts into a data frame with three columns:
documents, terms, counts. Each row is a unique document by term
count This is akin to \code{reshape2} packages \code{melt} function,
but works on a sparse matrix.The resulting data frame is
equivalent to the \code{tidytext} triplet tibble.
}
\author{
Dustin Stoltz
}
