% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ft_wv_sample}
\alias{ft_wv_sample}
\title{Sample of fastText embeddings}
\format{
A matrix of 770 rows and 300 columns
}
\usage{
ft_wv_sample
}
\description{
These are a sample of the English fastText embeddings
including 770 words matching those used in the \code{jfk_speech}.
These are intended to be used for example code.
}
\keyword{datasets}
