% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jfk_speech}
\alias{jfk_speech}
\title{Full Text of JFK's Rice Speech}
\format{
A data frame with 2 columns
}
\usage{
jfk_speech
}
\description{
This is a data frame for the text of JFK's Rice Speech "We choose to go to
the moon." Each row is a 10 word string of the speech -- roughly a sentence.
This is intended to be used for example code.
}
\section{Variables}{

Variables:
\itemize{
\item sentence_id. Order and unique ID for the sentence
\item sentence. The text of a sentence
}
}

\keyword{datasets}
