% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_shakespeare}
\alias{meta_shakespeare}
\title{Metadata for Shakespeare's First Folio}
\format{
A matrix of 37 rows and 8 columns
}
\usage{
meta_shakespeare
}
\description{
Metadata related to Shakespeare's First Folio
including the IDs to download the plays from
Project Gutenberg, and a count of the number of
deaths in each play (body count).
}
\section{Variables}{

Variables:
\itemize{
\item short_title.
\item gutenberg_title.
\item gutenberg_id.
\item genre.
\item year.
\item body_count.
\item boas_problem_plays.
\item death.
}
}

\keyword{datasets}
