% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoCA.R
\name{plot.CoCA}
\alias{plot.CoCA}
\title{Plot CoCA}
\usage{
\method{plot}{CoCA}(
  x,
  module = NULL,
  cutoff = 0.05,
  repulse = 1.86,
  min = 0.15,
  max = 1,
  main = NULL,
  ...
)
}
\arguments{
\item{x}{CoCA object returned by \code{\link[=CoCA]{CoCA()}}}

\item{module}{index for which module to plot (default = NULL)}

\item{cutoff}{minimum absolute value of correlations to plot}

\item{repulse}{repulse radius in the spring layout}

\item{min}{edges with absolute weights under this value are
not shown (default = 0.15)}

\item{max}{highest weight to scale the edge widths too (default = 1)}

\item{main}{title for plot (default = NULL)}

\item{...}{Arguments to be passed to methods}
}
\value{
returns \code{qgraph} object
}
\description{
Plot CoCA
}
