% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoCA.R
\name{print.CoCA}
\alias{print.CoCA}
\title{Prints CoCA class information}
\usage{
\method{print}{CoCA}(x, ...)
}
\arguments{
\item{x}{CoCA object returned by \code{CoCA()}}

\item{...}{Arguments to be passed to methods}
}
\value{
prints a message indicating the classes and sizes
}
\description{
Prints CoCA class information
}
