% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2map-package.R
\docType{package}
\name{text2map-package}
\alias{text2map}
\alias{text2map-package}
\title{Text2Map}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This is a collection of functions optimized for working with with various kinds of text matrices. Focusing on the text matrix as the primary object - represented either as a base R dense matrix or a 'Matrix' package sparse matrix - allows for a consistent and intuitive interface that stays close to the underlying mathematical foundation of computational text analysis. In particular, the package includes functions for working with word embeddings, text networks, and document-term matrices. Methods developed in Stoltz and Taylor (2019) <doi:10.1007/s42001-019-00048-6>, Taylor and Stoltz (2020) <doi:10.1007/s42001-020-00075-8>, Taylor and Stoltz (2020) <doi:10.15195/v7.a23>, and Stoltz and Taylor (2021) <doi:10.1016/j.poetic.2021.101567>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/culturalcartography/text2map}
  \item Report bugs at \url{https://gitlab.com/culturalcartography/text2map/-/issues}
}

}
\author{
\strong{Maintainer}: Dustin Stoltz \email{dss219@lehigh.edu} (\href{https://orcid.org/0000-0002-4774-0765}{ORCID})

Authors:
\itemize{
  \item Marshall Taylor \email{mtaylor2@nmsu.edu} (\href{https://orcid.org/0000-0002-7440-0723}{ORCID})
}

}
\keyword{internal}
