% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tiny_gender_tagger}
\alias{tiny_gender_tagger}
\title{A very tiny "gender" tagger}
\usage{
tiny_gender_tagger()
}
\value{
returns a tibble with two columns
}
\description{
Provides a small dictionary which matches common English pronouns
and nouns to conventional gender categories ("masculine" or
"feminine"). There are 20 words in each category.
}
\author{
Dustin Stoltz
}
