% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anchor_lists}
\alias{anchor_lists}
\title{A dataset of anchor lists}
\format{
A data frame with 303 rows and 4 variables.
}
\usage{
anchor_lists
}
\description{
A dataset containing juxtaposing pairs of English words
for 26 semantic relations. These anchors are used with
the \code{get_anchor()} function, which can then be used with
the \code{get_direction()} function.
}
\section{Variables}{

Variables:
\itemize{
\item add. words to added (or the positive direction)
\item subtract. words to be subtract (or the negative direction)
\item relation. the relation to be extracted, 26 relations available
\item domain. 6 broader categories within which each relation falls

}
}

\keyword{datasets}
